automatically backed up by learn

+++ index.js
@@ -27,7 +27,7 @@ function dec(n) {
 }
 
 function makeInt(n) {
-  return parseFloat(n)
+  return parseInt(n)
 }
 
 function makeInt(n) {
